+++ dart/lib/src/exception_cause.dart
@@ -1,0 +1,6 @@
+class ExceptionCause {
+  ExceptionCause(this.exception, this.stackTrace);
+
+  dynamic exception;
+  dynamic stackTrace;
+}